% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFromLm.R
\name{createFromLm}
\alias{createFromLm}
\title{Create a Flexible Transfer Model (ftmlm) from a lm object}
\usage{
createFromLm(lmObj)
}
\arguments{
\item{lmObj}{a lm object from the stats package}
}
\value{
A ftmlm object.
}
\description{
This function extracts necessary components from a lm object to create an ftmlm object.
It extracts the necessary components and adapts them into the FTM framework.
}
\examples{
\dontrun{
# Fitting a linear model to the mtcars dataset
data(mtcars)
lm_model <- lm(mpg ~ cyl + hp + wt, data = mtcars)
ftmlm_model <- createFromLm(lm_model)
}
}
