% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{formula,ftmglm-method}
\alias{formula,ftmglm-method}
\alias{formula,ftmlm-method}
\title{Model Formula for ftmglm or ftmlm objects}
\usage{
\S4method{formula}{ftmglm}(x)

\S4method{formula}{ftmlm}(x)
}
\arguments{
\item{x}{Object of class inheriting from \code{ftmglm} or \code{ftmlm}.}
}
\value{
an object of class 'formula'
}
\description{
Retrieves the formula used to create a \code{ftmglm} or \code{ftmlm} model
}
\note{
If the \code{ftmglm} or \code{ftmlm} object was created from a \code{cv.glmnet} object,
the outcome variable will be named "y", rather than the original name.
}
