% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftmglm.R
\docType{class}
\name{ftmglm-class}
\alias{ftmglm-class}
\title{Flexible Transfer Model (FTM) Generalized Linear Model Class}
\description{
An S4 class designed to encapsulate the necessary components for a flexible
transfer model generalized linear model. This class stores matrices that
represent the weighted cross-product of predictors and the cross-product of predictors
with outcomes, essential for the model calculations.
}
\section{Usage}{

This class is internally used by functions within the package to perform
regression analysis. It is not intended for direct use.
}

\section{Slots}{

\describe{
  \item{XtWX}{\code{matrix} representing the weighted cross product of the predictor variables.}
  \item{XtWz}{\code{matrix} representing the weighted cross product of the predictor variables and outcomes.}
}
}

