% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftmglm.R
\name{ftmglm}
\alias{ftmglm}
\title{Constructor for ftmglm Objects}
\usage{
ftmglm(XtWX, XtWz)
}
\arguments{
\item{XtWX}{A numeric matrix representing the weighted crossproduct of predictors.}

\item{XtWz}{A numeric matrix representing the weighted crossproduct of predictors and outcomes.}
}
\value{
An object of class \code{\linkS4class{ftmglm}}.
}
\description{
Creates an instance of the \code{\linkS4class{ftmglm}} class, initializing
it with specific matrices necessary for the flexible transfer model calculations.
This function is typically called internally within higher-level functions that
prepare data and perform model fitting.
}
\examples{
\dontrun{
  XtWX <- matrix(rnorm(100), 10, 10)
  XtWz <- matrix(rnorm(10), 10, 1)
  model <- ftmglm(XtWX, XtWz)
}
}
\seealso{
\code{\link{ftmglm-class}} for details on the class structure.
}
