% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftmlm.R
\docType{class}
\name{ftmlm-class}
\alias{ftmlm-class}
\title{Flexible Transfer Model (FTM) Linear Model Class}
\description{
An S4 class designed to encapsulate the necessary components for a flexible
transfer model linear model. This class stores matrices that
represent the weighted cross-product of predictors and the cross-product of predictors
with outcomes, essential for the model calculations.
}
\section{Usage}{

This class is internally used by functions within the package to perform
regression analysis. It is not intended for direct use.
}

\section{Slots}{

\describe{
  \item{XtX}{\code{matrix} representing the crossproduct of the predictor variables.}
  \item{Xty}{\code{matrix} representing the crossproduct of the predictor variables and outcomes.}
  \item{s}{\code{numeric} representing the optimal Ridge penalty from cross-validation (if available).}
}
}

