% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsq.R
\name{rsq}
\alias{rsq}
\title{Extract R-squared from model objects}
\usage{
rsq(object, ...)
}
\arguments{
\item{object}{A model object from which to extract the R-squared value}

\item{...}{Additional arguments passed to methods}
}
\value{
A numeric value representing the R-squared or pseudo-R-squared value
}
\description{
Generic function to extract R-squared values from various model types
in the FTM package.
}
