% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFromGlm.R
\name{createFromGlm}
\alias{createFromGlm}
\title{Create a Flexible Transfer Model (ftmglm or ftmlm) from a glm object}
\usage{
createFromGlm(glmObj)
}
\arguments{
\item{glmObj}{a glm object from the stats package}
}
\value{
A ftmglm object.
}
\description{
This function extracts necessary components from a glm object to create an ftmglm object.
It extracts the necessary components and adapts them into the FTM framework.
}
\examples{
\dontrun{
# Fitting a glm to the mtcars dataset
data(mtcars)
glm_model <- glm(am ~ hp + wt + cyl, data = mtcars, family = "binomial")
ftmglm_model <- createFromGlm(glm_model)
}
}
