% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftmlm.R
\name{ftmlm}
\alias{ftmlm}
\title{Constructor for ftmlm Objects}
\usage{
ftmlm(XtX, Xty, s = NULL)
}
\arguments{
\item{XtX}{A numeric matrix representing the crossproduct of predictors.}

\item{Xty}{A numeric matrix representing the crossproduct of predictors and outcomes.}

\item{s}{A numeric value representing the optimal Ridge penalty from cross-validation (if available).}
}
\value{
An object of class \code{\linkS4class{ftmlm}}.
}
\description{
Creates an instance of the \code{\linkS4class{ftmlm}} class, initializing
it with specific matrices necessary for the flexible transfer model calculations.
This function is typically called internally within higher-level functions that
prepare data and perform model fitting.
}
\examples{
\dontrun{
  XtX <- matrix(rnorm(100), 10, 10)
  Xty <- matrix(rnorm(10), 10, 1)
  model <- ftmlm(XtX, Xty)
}
}
\seealso{
\code{\link{ftmlm-class}} for details on the class structure.
}
